requirements.txt
pyqt5==5.11.3
asyncqt==0.7.0
aiofiles==0.4.0
attrs==18.2.0
pypiwin32==223 ; sys_platform == 'win32'
pygobject==3.30.4 ; sys_platform == 'linux'
aiohttp==3.5.4
google-api-python-client==1.7.7
oauth2client==4.1.3
facebook-sdk==3.1.0
furl==2.0.0
